(WIP) fix ui tests
