+++ test/source/test.ts
@@ -23,7 +23,7 @@ import { TestUrls } from './browser/test-urls';
 import { mkdirSync, realpathSync, writeFileSync } from 'fs';
 import { startAllApisMock } from './mock/all-apis-mock';
 import { defineContentScriptTests } from './tests/content-script';
-import { LEGACY_GLOBAL_ATTESTER_MOCK_CONFIG } from './mock/attester/attester-endpoints';
+import { ATTESTER_ACCOUNT_INIT_SETUP_MOCK_CONFIG } from './mock/attester/attester-key-contstants';
 
 export const { testVariant, testGroup, oneIfNotPooled, buildDir, isMock } = getParsedCliParams();
 export const internalTestState = { expectIntentionalErrReport: false }; // updated when a particular test that causes an error is run
@@ -68,7 +68,7 @@ const testWithBrowser = (
     let closeMockApi: (() => Promise<void>) | undefined;
     if (isMock) {
       t.mockApi = await startMockApiAndCopyBuild(t);
-      t.mockApi.attesterConfig = LEGACY_GLOBAL_ATTESTER_MOCK_CONFIG;
+      t.mockApi.attesterConfig = ATTESTER_ACCOUNT_INIT_SETUP_MOCK_CONFIG;
       closeMockApi = t.mockApi.close;
     } else {
       t.urls = new TestUrls(await browserPool.getExtensionId(t));
