+++ test/source/tests/settings.ts
@@ -25,6 +25,7 @@ import { AvaContext } from './tooling';
 import { mockBackendData } from '../mock/backend/backend-endpoints';
 import { ClientConfiguration, keyManagerAutogenRules } from '../mock/backend/backend-data';
 import { HttpClientErr, Status } from '../mock/lib/api';
+import { somePubkey, testMatchPubKey } from '../mock/attester/attester-key-contstants';
 
 export const defineSettingsTests = (testVariant: TestVariant, testWithBrowser: TestWithBrowser) => {
   if (testVariant !== 'CONSUMER-LIVE-GMAIL') {
@@ -62,6 +63,16 @@ export const defineSettingsTests = (testVariant: TestVariant, testWithBrowser: T
       testWithBrowser(undefined, async (t, browser) => {
         const email = 'test.match.attester.key@gmail.com';
         const mismatchEmail = 'test.mismatch.attester.key@gmail.com';
+        t.mockApi!.attesterConfig = {
+          pubkeyLookup: {
+            [email]: {
+              pubkey: testMatchPubKey,
+            },
+            [mismatchEmail]: {
+              pubkey: somePubkey,
+            },
+          },
+        };
         const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, email);
         await SetupPageRecipe.manualEnter(
           settingsPage,
