+++ test/source/tests/setup.ts
@@ -10,7 +10,6 @@ import { SettingsPageRecipe } from './page-recipe/settings-page-recipe';
 import { ComposePageRecipe } from './page-recipe/compose-page-recipe';
 import { Str, emailKeyIndex } from './../core/common';
 import { MOCK_KM_LAST_INSERTED_KEY, MOCK_KM_KEYS } from './../mock/key-manager/key-manager-endpoints';
-import { MOCK_ATTESTER_LAST_INSERTED_PUB, hasPubKey, protonMailCompatKey } from './../mock/attester/attester-endpoints';
 import { BrowserRecipe } from './tooling/browser-recipe';
 import { Key, KeyInfoWithIdentity, KeyUtil } from '../core/crypto/key';
 import { testConstants } from './tooling/consts';
@@ -20,6 +19,8 @@ import { BrowserHandle, ControllablePage } from '../browser';
 import { OauthPageRecipe } from './page-recipe/oauth-page-recipe';
 import { AvaContext } from './tooling';
 import { opgp } from '../core/crypto/pgp/openpgpjs-custom';
+import { MOCK_ATTESTER_LAST_INSERTED_PUB } from '../mock/attester/attester-endpoints';
+import { hasPubKey, protonMailCompatKey, somePubkey } from '../mock/attester/attester-key-contstants';
 
 const getAuthorizationHeader = async (t: AvaContext, browser: BrowserHandle, acctEmail: string) => {
   const settingsPage = await browser.newExtensionSettingsPage(t, acctEmail);
@@ -802,14 +803,15 @@ AN8G3r5Htj8olot+jm9mIa5XLXWzMNUZgg==
     test(
       'has.pub@client-configuration-test.flowcrypt.test - no backup, no keygen',
       testWithBrowser(undefined, async (t, browser) => {
+        const acct = 'has.pub@client-configuration-test.flowcrypt.test';
         t.mockApi!.attesterConfig = {
           ldapRelay: {
-            'has.pub@client-configuration-test.flowcrypt.test': {
+            [acct]: {
               pubkey: hasPubKey,
             },
           },
         };
-        const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, 'has.pub@client-configuration-test.flowcrypt.test');
+        const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, acct);
         await SetupPageRecipe.manualEnter(
           settingsPage,
           'has.pub.client.configuration.test',
@@ -836,14 +838,15 @@ AN8G3r5Htj8olot+jm9mIa5XLXWzMNUZgg==
     test(
       'invalid.pub@client-configuration-test.flowcrypt.test - no backup, no keygen',
       testWithBrowser(undefined, async (t, browser) => {
+        const acct = 'invalid.pub@client-configuration-test.flowcrypt.test';
         t.mockApi!.attesterConfig = {
           ldapRelay: {
-            'invalid.pub@client-configuration-test.flowcrypt.test': {
+            [acct]: {
               pubkey: protonMailCompatKey,
             },
           },
         };
-        const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, 'invalid.pub@client-configuration-test.flowcrypt.test');
+        const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, acct);
         await SetupPageRecipe.manualEnter(
           settingsPage,
           'has.pub.client.configuration.test',
@@ -929,6 +932,16 @@ AN8G3r5Htj8olot+jm9mIa5XLXWzMNUZgg==
       testWithBrowser(undefined, async (t, browser) => {
         // disallowed searching attester for pubkeys on "flowcrypt.com" domain
         // below we search for human@flowcrypt.com which normally has pubkey on attester, but none should be found due to the rule
+        t.mockApi!.attesterConfig = {
+          pubkeyLookup: {
+            'mock.only.pubkey@flowcrypt.com': {
+              pubkey: somePubkey,
+            },
+            'mock.only.pubkey@other.com': {
+              pubkey: somePubkey,
+            },
+          },
+        };
         const acct = 'user@no-search-domains-client-configuration.flowcrypt.test';
         const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, acct);
         await SetupPageRecipe.manualEnter(settingsPage, 'flowcrypt.test.key.used.pgp');
@@ -949,6 +962,16 @@ AN8G3r5Htj8olot+jm9mIa5XLXWzMNUZgg==
       testWithBrowser(undefined, async (t, browser) => {
         // disallow_attester_search_for_domains is not respected if allow_attester_search_only_for_domains is set
         // searching attester for pubkeys only on "flowcrypt.com" domain
+        t.mockApi!.attesterConfig = {
+          pubkeyLookup: {
+            'mock.only.pubkey@flowcrypt.com': {
+              pubkey: somePubkey,
+            },
+            'mock.only.pubkey@other.com': {
+              pubkey: somePubkey,
+            },
+          },
+        };
         const acct = 'user@only-allow-some-domains-client-configuration.flowcrypt.test';
         const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, acct);
         await SetupPageRecipe.manualEnter(settingsPage, 'flowcrypt.test.key.used.pgp');
